Fix bugs
- Talbe pay bugs: Only show the unpaid table
- The quantity should be real number not integer

+++ app/src/main/java/com/giao/ordersystem/OrderDetailsBO.java
@@ -7,14 +7,14 @@ public class OrderDetailsBO {
     private int orderDetailID;
     private int orderID;
     private int dishID;
-    private int quantity;
+    private Float quantity;
     private Float price;
     private String note;
     public OrderDetailsBO()
     {
 
     }
-    public OrderDetailsBO(int orderDetailID, int orderID, int dishID,int quantity, Float price, String note)
+    public OrderDetailsBO(int orderDetailID, int orderID, int dishID,Float quantity, Float price, String note)
     {
         this.orderDetailID=orderDetailID;
         this.orderID=orderID;
@@ -29,8 +29,8 @@ public class OrderDetailsBO {
     public void setOrderID(int orderID) {this.orderID = orderID;}
     public int getDishID(){return this.dishID;}
     public void setDishID(int note){this.dishID=dishID;}
-    public int getQuantity(){return this.quantity;}
-    public void setQuantity(int quantity){this.quantity=quantity;}
+    public Float getQuantity(){return this.quantity;}
+    public void setQuantity(Float quantity){this.quantity=quantity;}
     public Float getPrice(){return this.price;}
     public void setPrice(Float price){this.price=price;}
     public String getNote(){return this.note;}
